pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inclusion-pipeline"
version = "0.1.0"
description = "MicroSAM segmentation + ResNet101 classification pipeline for microscopy images"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "sunghoojung"}
]
keywords = ["segmentation", "classification", "microsam", "deep-learning", "microscopy", "inclusion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
    "scikit-image>=0.21.0",
    "huggingface_hub>=0.16.0",
    # micro_sam must be installed separately via conda:
    # conda install -c conda-forge micro_sam
]

[project.urls]
Homepage = "https://github.com/sunghoojung/segment-classify-pipeline"
Repository = "https://github.com/sunghoojung/segment-classify-pipeline"

[tool.setuptools.packages.find]
include = ["inclusion_pipeline*"]
